Removing wandb.init() call when sweeping
We need a call to wandb.init() outside of the train() method anyway when starting a sweep in any case to obtain the sweep parameters. Then, the inner call to wandb.init() crashes the sweep with wandb 0.10.10 and 0.10.11 for me. So I only do the inner call when there is no sweep active determined by the passing of a sweep_config.